Break cyclic imports.
