+++ revelation/execute_branch.py
@@ -1,7 +1,7 @@
 from revelation.condition_codes import condition_passed
+from revelation.registers import reg_map
 from revelation.utils import signed, sext_8, sext_24, trim_32
 
-import revelation.isa
 
 def make_bcond_executor(is16bit):
     def execute_bcond(s, inst):
@@ -18,7 +18,7 @@ def make_bcond_executor(is16bit):
         cond = inst.cond
         imm = inst.bcond_imm
         if cond == 0b1111:  # Branch and link (BL).
-            s.rf[revelation.isa.reg_map['LR']] = s.pc + (2 if is16bit else 4)
+            s.rf[reg_map['LR']] = s.pc + (2 if is16bit else 4)
         if condition_passed(s, cond):
             offset = (signed(sext_8(imm)) << 1) if is16bit else (signed(sext_24(imm)) << 1)
             s.pc = trim_32(s.pc + offset)
